Vrai finalisation
Optimisations faites pour les boucles + ajout de const

+++ src/Board.cpp
@@ -17,8 +17,8 @@ std::array<std::array<char, 3>, 3> createBoard() {
     std::array<std::array<char, 3>, 3> Board;
 
     char value = '1';
-    for (int i {0}; i < Board.size(); ++i) {
-        for (int j {0}; j < Board.size(); ++j) {
+    for (size_t i {0}; i < Board.size(); ++i) {
+        for (size_t j {0}; j < Board.size(); ++j) {
             Board[i][j] = value++;
         }
     }
